fix: ensure stages are ordered (#163)
* fix: ensure stages are ordered

Stage ordering has been relying on absolute ordering of stage indexes
within a game. Also stage indexes were applied on callback finalizers in
a loop where multiple stages on a round could acquire their stage index
out of order.

Stages are now ordered within a round. This is done so that stages can
be added to rounds out of order (add stage 1 to round 1, stage 1 to
round 2 and stage 2 to round 1 ; this scenario would have made the
stages go out of order). Also, rounds are now ordereds. Finally, all
stages added at the same time as a round (within the same callback)
should not be ordered.

* fix: make round and stage indexes 0-indexed

* feat: upgrade tajriba to v1.0.0-rc.17